populated user in posts

+++ app/routers/post.py
@@ -1,4 +1,4 @@
-from .. import models, schemas
+from .. import models, schemas, oauth2
 from fastapi import Response, status, HTTPException, Depends, APIRouter
 from ..database import get_db
 from sqlalchemy.orm import Session
@@ -10,11 +10,11 @@ router = APIRouter(
 )
 
 @router.post("/", status_code=status.HTTP_201_CREATED, response_model=schemas.PostResponse)
-async def create_posts(post: schemas.PostCreate, db: Session = Depends(get_db)):
+async def create_posts(post: schemas.PostCreate, db: Session = Depends(get_db), current_user: int = Depends(oauth2.get_current_user)):
     # cursor.execute("""INSERT INTO posts (title, content, published) VALUES (%s, %s, %s) RETURNING * """, (post.title, post.content, post.published))
     # new_post = cursor.fetchone()
     # conn.commit()
-    new_post = models.Post(**post.model_dump())
+    new_post = models.Post(owner_id=current_user.id, **post.model_dump())
     db.add(new_post)
     db.commit()
     db.refresh(new_post)
@@ -27,9 +27,22 @@ async def get_posts(db: Session = Depends(get_db)):
     posts = db.query(models.Post).all()
     return  posts
 
+@router.get("/by-me", response_model=List[schemas.PostResponse])
+async def get_my_posts(
+    db: Session = Depends(get_db),
+    current_user: models.User = Depends(oauth2.get_current_user)
+):
+    posts = db.query(models.Post).filter(models.Post.owner_id == current_user.id).all()
+
+    if not posts:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail="No posts found for the current user"
+        )
+    return posts
 
 @router.get('/{id}',response_model=schemas.PostResponse)
-async def get_post(id: int, db: Session = Depends(get_db)):
+async def get_post(id: int, db: Session = Depends(get_db), current_user: int = Depends(oauth2.get_current_user)):
     # cursor.execute("""SELECT * FROM posts WHERE id = %s""", (str(id),))
     # post = cursor.fetchone()
     post = db.query(models.Post).filter(models.Post.id == id).first()
@@ -39,19 +52,21 @@ async def get_post(id: int, db: Session = Depends(get_db)):
 
 
 @router.delete('/{id}')
-async def delete_post(id: int, db:Session = Depends(get_db)):
+async def delete_post(id: int, db:Session = Depends(get_db), current_user: int = Depends(oauth2.get_current_user)):
     # cursor.execute("""DELETE FROM posts WHERE id = %s RETURNING *""", (str(id),))
     # post = cursor.fetchone()
     # conn.commit()
     post = db.query(models.Post).filter(models.Post.id == id).first()
     if not post:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=f"post with id: {id} was not found")
+    if post.owner_id != current_user.id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail=f"not authorized to perform requested action")
     db.delete(post)
     db.commit()
     return Response(status_code = status.HTTP_204_NO_CONTENT)
 
 @router.put('/{id}', response_model=schemas.PostResponse)
-async def update_post(id: int, updated_post: schemas.PostCreate, db: Session = Depends(get_db)):
+async def update_post(id: int, updated_post: schemas.PostCreate, db: Session = Depends(get_db), current_user: int = Depends(oauth2.get_current_user)):
     # cursor.execute("""UPDATE posts SET title = %s, content = %s, published = %s WHERE id = %s RETURNING *""", (post.title, post.content, post.published, str(id)))
     # post = cursor.fetchone()
     # conn.commit()
@@ -59,5 +74,7 @@ async def update_post(id: int, updated_post: schemas.PostCreate, db: Session = D
     post = post_query.first()
     if not post:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=f"post with id: {id} was not found")
+    if post.owner_id != current_user.id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail=f"not authorized to perform requested action")
     post_query.update(updated_post.model_dump(), synchronize_session=False)
     return post_query.first()
